requirements.txt
# API and HTTP
requests>=2.28.1
python-dotenv>=0.21.0

# Audio processing
pydub>=0.25.1

# File handling and utilities
pillow>=9.3.0

# Progress tracking and CLI
tqdm>=4.64.1

# JSON handling
jsonschema>=4.17.3 